Moving all app functions to support + removing app.obj from realmOfFeastTest
